+++ core/user_db.service.js
@@ -72,12 +72,11 @@ async function create_db_request(asked_db, user) {
             {
                 name: 'ask_database',
                 destinations: msg_destinations,
-                subject: 'Database creation request: ' + asked_db.id
+                subject: 'Database creation request: ' + asked_db.name
             },
             {
                 '#UID#': user.uid,
                 '#NAME#': asked_db.name,
-                '#SIZE#': asked_db.size,
                 '#USAGE#': asked_db.usage,
                 '#EXP#': asked_db.expire
             }